# DS-tuts/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Online Order API",
    "version": "1.0.0",
    "description": "A basic API for working with Swagger tools"
  },
  "paths": {
    "/orders": {
      "summary": "GET all of the orders data",
      "description": "This path is used to retrieve all of the orders data from the orders.json file",
      "get": {
        "summary": "Gets the order data",
        "description": "Retrieve the order information from the orders.ison file",
        "operationId": "get_orders",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "examples": {
                  "orders": {
                    "value": "{\"orders\":[\n  {\n    \"name\": \"Carey Maynard\",\n    \"id\":\"001\",\n    \"state\": \"pending\"\n  },\n  {\n    \"name\": \"Angela Ayala\",\n    \"id\":\"002\",\n    \"state\": \"canceled\"\n  },\n  {\n    \"name\": \"Regina Yates\",\n    \"id\":\"003\",\n    \"state\": \"pending\"\n  }\n  ]}"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Default error sample response"
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/neworder": {
      "summary": "POST a new order",
      "description": "This path is used to ADD a new order to the orders.json file",
      "post": {
        "summary": "Add a new order",
        "description": "this path uses POST to add a new order to the orders.json file",
        "operationId": "post_orders",
        "requestBody": {
          "description": "A new order object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Order"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "Default error sample response"
          }
        },
        "tags": [
          "New Order"
        ]
      }
    },
    "/update/{id}": {
      "summary": "Change the status of an order",
      "description": "This path is used to change the status of an order matching the provided id. It modifies the state attribute of each order.",
      "put": {
        "summary": "Update the state of an order with matching id",
        "description": "This path uses PUT to update the state of an order with an matching in the orders.json file",
        "operationId": "put_orders",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the order.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Default error sample response"
          }
        },
        "tags": [
          "Update Order"
        ]
      }
    },
    "/delete/{id}": {
      "summary": "DELETE an order",
      "description": "This path used to delete an order with a matching id",
      "delete": {
        "summary": "Delete an order with matching id",
        "description": "This path DELETE's an order with a matching id from the order.json file",
        "operationId": "delete_orders",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the order.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Default error sample response"
          }
        },
        "tags": [
          "Delete Order"
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "Order": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "state": {
            "type": "string"
          }
        },
        "xml": {
          "name": "Order"
        }
      }
    }
  }
}